--- .env-sample
@@ -1,1 +1,0 @@
-user_id =<user-id-of-the-person>
