+++ .gitignore
@@ -10,7 +10,7 @@ lerna-debug.log*
 report.[0-9]*.[0-9]*.[0-9]*.[0-9]*.json
 
 #folders
-images
+
 
 # Runtime data
 pids
@@ -84,8 +84,6 @@ typings/
 # Nuxt.js build / generate output
 .nuxt
 dist
-images
-images-pm
 
 # Gatsby files
 .cache/
