+++ index.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset='utf-8'>
+    <meta http-equiv='X-UA-Compatible'>
+    <title>Page Title</title>
+    <meta name='viewport' content='width=device-width, initial-scale=1'>
+    <link rel='stylesheet' type='text/css' media='screen' href='enigma.css'>
+    <script defer src='enigma.js'></script>
+</head>
+<body>
+    <div id="div_wrapper">
+        <!-- <div id="div_picture"></div> -->
+        <div id="div_picture2" ></div>
+    </div>
+</body>
+
+</html>
