--- package.json
@@ -1,25 +1,0 @@
-{
-  "name": "shiny-enigma",
-  "version": "1.0.0",
-  "description": "",
-  "main": "new.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ashwinadiga01/shiny-enigma.git"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/ashwinadiga01/shiny-enigma/issues"
-  },
-  "homepage": "https://github.com/ashwinadiga01/shiny-enigma#readme",
-  "dependencies": {
-    "dotenv": "^10.0.0",
-    "node-fetch": "^2.6.1",
-    "request": "^2.88.2"
-  }
-}
