--- ui/enigma.css
@@ -1,30 +1,0 @@
-
-.static {
-    position:absolute;
-    background: white;
-  }
-  
-  .static:hover {
-    opacity:0;
-  }
-
-
-.div_picture1.img {
-    /* filter: url(filters.svg#grayscale); Firefox 3.5+ */
-    filter: grayscale(0%); /* IE5+ */
-    -webkit-filter: grayscale(0); /* Webkit Nightlies & Chrome Canary */
-    -webkit-transition: all .8s ease-in-out;  
-    }
-
-    img {
-    filter: grayscale(0%);
-    -webkit-filter: grayscale(1);
-    -webkit-transform: scale(1);
-    }
-
-    img:hover {
-        filter: grayscale(0%);
-        -webkit-filter: grayscale(0);
-        -webkit-transform: scale(1.01);
-    }
-
